// frontend/src/app/api/models/activity-summary.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ActivitySummary {
  best100Duration?: string;
  best13Duration?: string;
  best1Duration?: string;
  best25Duration?: string;
  best2Duration?: string;
  best4Duration?: string;
  best50Duration?: string;
  best8Duration?: string;
  chipId?: number;
  endTime?: string;
  id?: number;
  locationId?: number;
  name?: string;
  seasonId?: number;
  startTime?: string;
}